refactor: change BoardInfo fields from int to size_t
- Update FileLoader::BoardInfo struct fields (rows, cols, maxSteps, numShells) from int to size_t
- Enhance parseKeyValue to validate against negative values while accepting size_t
- Remove explicit static_cast operations in game_runner.cpp
- Update all test files and helper functions to use size_t types
- Add comprehensive negative number validation tests

Improves type safety and eliminates unnecessary casting between int and size_t.

+++ Simulator/game_runner.cpp
@@ -16,11 +16,10 @@ GameResult GameRunner::runSingleGame(
         throw std::runtime_error("Invalid board info: null satellite view");
     }
     
-    // TODO: cosider replacing board info to ints
-    size_t mapWidth = static_cast<size_t>(boardInfo.cols);
-    size_t mapHeight = static_cast<size_t>(boardInfo.rows);
-    size_t maxSteps = static_cast<size_t>(boardInfo.maxSteps);
-    size_t numShells = static_cast<size_t>(boardInfo.numShells);
+    size_t mapWidth = boardInfo.cols;
+    size_t mapHeight = boardInfo.rows;
+    size_t maxSteps = boardInfo.maxSteps;
+    size_t numShells = boardInfo.numShells;
     auto& satelliteView = *boardInfo.satelliteView;
     
     // 2. Get GameManager factory from registrar
